--- src/main/groovy/com/github/hauner/openapi/micronaut/writer/java/MappingAnnotationWriter.groovy
@@ -1,72 +1,0 @@
-/*
- * Copyright 2020 the original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.hauner.openapi.micronaut.writer.java
-
-import com.github.hauner.openapi.core.writer.java.MappingAnnotationWriter as CoreMappingAnnotationWriter
-import com.github.hauner.openapi.core.model.Endpoint
-import com.github.hauner.openapi.core.model.EndpointResponse
-
-/**
- * micronaut mapping annotation writer
- *
- * @author Martin Hauner
- */
-class MappingAnnotationWriter implements CoreMappingAnnotationWriter {
-
-    @Override
-    void write (Writer target, Endpoint endpoint, EndpointResponse endpointResponse) {
-        target.write (createAnnotation (endpoint, endpointResponse))
-    }
-
-    private String createAnnotation (Endpoint endpoint, EndpointResponse endpointResponse) {
-        String mapping = "${getMappingAnnotation (endpoint)}"
-        mapping += "("
-        mapping += 'uri = ' + quote(endpoint.path)
-
-        def consumes = endpoint.consumesContentTypes
-        if (!consumes.empty) {
-            mapping += ", "
-            mapping += 'consumes = {'
-            mapping +=  consumes.collect {
-                quote(it)
-            }.join (', ')
-            mapping += '}'
-        }
-
-        def produces = endpointResponse.contentTypes
-        if (!produces.empty) {
-            mapping += ", "
-            mapping += 'produces = {'
-            mapping += produces.collect {
-                quote (it)
-            }.join (', ')
-            mapping += '}'
-        }
-
-        mapping += ")"
-        mapping
-    }
-
-    private String getMappingAnnotation (Endpoint endpoint) {
-        "@${endpoint.method.method.capitalize ()}"
-    }
-
-    private String quote (String content) {
-        '"' + content + '"'
-    }
-
-}
