+++ src/main/kotlin/io/openapiprocessor/micronaut/writer/java/MappingAnnotationWriter.kt
@@ -1,0 +1,81 @@
+/*
+ * Copyright 2020 the original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.openapiprocessor.micronaut.writer.java
+
+import io.openapiprocessor.core.writer.java.MappingAnnotationWriter as CoreMappingAnnotationWriter
+import io.openapiprocessor.core.model.Endpoint
+import io.openapiprocessor.core.model.EndpointResponse
+import java.io.Writer
+
+/**
+ * micronaut mapping annotation writer
+ *
+ * @author Martin Hauner
+ */
+class MappingAnnotationWriter: CoreMappingAnnotationWriter {
+
+    override fun write(target: Writer, endpoint: Endpoint, endpointResponse: EndpointResponse) {
+        target.write(createAnnotation(endpoint, endpointResponse))
+    }
+
+    private fun createAnnotation(endpoint: Endpoint, endpointResponse: EndpointResponse): String {
+        var mapping = getMappingAnnotation(endpoint)
+        mapping += "("
+        mapping += "uri = " + quote(endpoint.path)
+
+        if (endpoint.requestBodies.isNotEmpty()) {
+            mapping += ", "
+            mapping += "consumes = {" + quote(endpoint.getRequestBody().contentType) + "}"
+        }
+
+        val contentTypes = endpointResponse.contentTypes
+        if (contentTypes.isNotEmpty()) {
+            mapping += ", "
+            mapping += "produces = {"
+
+            mapping += contentTypes.map {
+                quote (it)
+            }.joinToString (", ")
+
+            mapping += "}"
+        }
+
+        /*
+        def consumes = endpoint.consumesContentTypes
+        if (!consumes.empty) {
+            mapping += ", "
+            mapping += 'consumes = {'
+            mapping +=  consumes.collect {
+                quote(it)
+            }.join (', ')
+            mapping += '}'
+        }*/
+
+        mapping += ")"
+        return mapping
+    }
+
+
+    private fun getMappingAnnotation(endpoint: Endpoint): String {
+        return "@${endpoint.method.method.capitalize ()}"
+    }
+
+    private fun quote(content: String): String {
+        return '"' + content + '"'
+    }
+
+}
